# docs/openapi.yaml
openapi: 3.1.0
info:
  title: RecallAI
  version: 0.1.0
paths:
...
  /api/v1/vault/uploads/limits:
    get:
      tags:
      - vault
      summary: Upload Limits
      description: 'What the client may send, so the picker can filter before a 25MB round-trip.


        This is deployment configuration -- it changes when the server restarts, never

        between two requests -- but the picker asks for it on every mount, and against a

        database in another region that was a ~600ms wait before the user could choose a

        file. `private` keeps it out of shared caches (it is behind a session cookie and

        describes that user''s own limits); the max-age is short enough that a config change

        is picked up within a minute of a deploy.'
      operationId: upload_limits_api_v1_vault_uploads_limits_get
      parameters:
      - name: recall_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Recall Session
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Upload Limits Api V1 Vault Uploads Limits Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/spaces:
    get:
      tags:
      - spaces
      summary: List Spaces
      operationId: list_spaces_api_v1_spaces_get
      parameters:
      - name: recall_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Recall Session
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SpaceRead'
                title: Response List Spaces Api V1 Spaces Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - spaces
      summary: Create Space
      operationId: create_space_api_v1_spaces_post
      parameters:
      - name: recall_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Recall Session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSpaceRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpaceRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/spaces/for-item/{item_id}:
    get:
      tags:
      - spaces
      summary: Spaces For Item
      description: 'Which of the caller''s Spaces already contain this memory.


        Drives the checkmarks in a card''s "+" menu. Answers with ids only, and only for

        Spaces the caller is in, so it cannot be used to probe whether a stranger filed

        something. An id the caller cannot see is simply absent -- there is no 404 here

        because "this memory is in none of your Spaces" is a real answer.


        Declared above the `/{space_id}` routes for the same reason the invite route is: a

        literal segment that could be read as a parameter has to be matched first.'
      operationId: spaces_for_item_api_v1_spaces_for_item__item_id__get
      parameters:
      - name: item_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Item Id
      - name: recall_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Recall Session
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                  format: uuid
                title: Response Spaces For Item Api V1 Spaces For Item  Item Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/spaces/invites/{invite_token}/accept:
    post:
      tags:
      - spaces
      summary: Accept Invite
      description: 'Spend an invite link.


        Every rejection is the same 404 -- unknown token, spent token, expired token, deleted

        Space. Distinguishing them tells whoever found the link in a screenshot exactly what

        they are holding.


        Declared **above** the `/{space_id}` routes, like `GET /vault/uploads/limits` is: a

        literal segment that could be read as a parameter has to be matched first. The path

        parameter is `invite_token` rather than `token` because FastAPI resolves path-parameter

        names across the whole dependency tree, and `get_current_user` already takes a `token`

        -- from a cookie, with a default, which a path parameter may not have.'
      operationId: accept_invite_api_v1_spaces_invites__invite_token__accept_post
      parameters:
      - name: invite_token
        in: path
        required: true
        schema:
          type: string
          title: Invite Token
      - name: recall_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Recall Session
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpaceRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/spaces/{space_id}:
    get:
      tags:
      - spaces
      summary: Get Space
      operationId: get_space_api_v1_spaces__space_id__get
      parameters:
      - name: space_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Space Id
      - name: recall_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Recall Session
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpaceDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - spaces
      summary: Update Space
      operationId: update_space_api_v1_spaces__space_id__patch
      parameters:
      - name: space_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Space Id
      - name: recall_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Recall Session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSpaceRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpaceRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - spaces
      summary: Delete Space
      operationId: delete_space_api_v1_spaces__space_id__delete
      parameters:
      - name: space_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Space Id
      - name: recall_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Recall Session
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/spaces/{space_id}/items:
    post:
      tags:
      - spaces
      summary: Add Items
      operationId: add_items_api_v1_spaces__space_id__items_post
      parameters:
      - name: space_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Space Id
      - name: recall_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Recall Session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddItemsRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddItemsResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/spaces/{space_id}/items/{item_id}:
    delete:
      tags:
      - spaces
      summary: Remove Item
      operationId: remove_item_api_v1_spaces__space_id__items__item_id__delete
      parameters:
      - name: space_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Space Id
      - name: item_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Item Id
      - name: recall_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Recall Session
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/spaces/{space_id}/connections:
    get:
      tags:
      - spaces
      summary: Space Connections
      description: 'The caller''s own connections between memories in this Space.


        **Only their own.** A Space is where one person reads another''s rows, and this does

        not widen that: a connection is a judgement its author made about their own memories,

        so two members see the same Space with different graphs. Widening later is a decision

        somebody can take deliberately; un-showing people each other''s judgements is not.


        Two statements: the membership gate, then the edges. Both ends must be in the Space

        and both must still exist, so nothing here can render a card from outside it.'
      operationId: space_connections_api_v1_spaces__space_id__connections_get
      parameters:
      - name: space_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Space Id
      - name: recall_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Recall Session
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpaceConnectionsResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/spaces/{space_id}/members:
    get:
      tags:
      - spaces
      summary: List Members
      operationId: list_members_api_v1_spaces__space_id__members_get
      parameters:
      - name: space_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Space Id
      - name: recall_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Recall Session
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SpaceMemberRead'
                title: Response List Members Api V1 Spaces  Space Id  Members Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/spaces/{space_id}/invites:
    post:
      tags:
      - spaces
      summary: Create Invite
      operationId: create_invite_api_v1_spaces__space_id__invites_post
      parameters:
      - name: space_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Space Id
      - name: recall_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Recall Session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInviteRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InviteResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/spaces/{space_id}/members/{member_id}:
    delete:
      tags:
      - spaces
      summary: Remove Member
      operationId: remove_member_api_v1_spaces__space_id__members__member_id__delete
      parameters:
      - name: space_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Space Id
      - name: member_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Member Id
      - name: recall_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Recall Session
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    AddItemsRequest:
      properties:
        item_ids:
          items:
            type: string
            format: uuid
          type: array
          maxItems: 100
          minItems: 1
          title: Item Ids
      type: object
      required:
      - item_ids
      title: AddItemsRequest
    AddItemsResponse:
      properties:
        added:
          type: integer
          title: Added
        skipped:
          type: integer
          title: Skipped
      type: object
      required:
      - added
      - skipped
      title: AddItemsResponse
      description: '`skipped` counts what was already there or was not the caller''s to add.


        Reported rather than raised: re-adding a memory is the normal outcome of "add all

        suggestions", not an error, and the old behaviour -- a primary-key violation surfacing

        as a 500 -- made the obvious user action look like a broken server.'
...
    CreateSpaceRequest:
      properties:
        name:
          type: string
          maxLength: 255
          minLength: 1
          title: Name
        description:
          anyOf:
          - type: string
            maxLength: 2000
          - type: 'null'
          title: Description
        visibility:
          $ref: '#/components/schemas/Visibility'
          default: private
        icon:
          anyOf:
          - type: string
            maxLength: 48
            pattern: ^[a-z][a-z0-9]*(-[a-z0-9]+)*$
          - type: 'null'
          title: Icon
        emoji:
          anyOf:
          - type: string
            maxLength: 16
          - type: 'null'
          title: Emoji
        accent:
          anyOf:
          - type: string
            maxLength: 24
          - type: 'null'
          title: Accent
        item_ids:
          items:
            type: string
            format: uuid
          type: array
          maxItems: 100
          title: Item Ids
      type: object
      required:
      - name
      title: CreateSpaceRequest
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    SpaceDetail:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        name:
          type: string
          title: Name
        slug:
          type: string
          title: Slug
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        visibility:
          $ref: '#/components/schemas/Visibility'
        icon:
          anyOf:
          - type: string
          - type: 'null'
          title: Icon
        emoji:
          anyOf:
          - type: string
          - type: 'null'
          title: Emoji
        accent:
          anyOf:
          - type: string
          - type: 'null'
          title: Accent
        pinned:
          type: boolean
          title: Pinned
        ai_overview:
          anyOf:
          - type: string
          - type: 'null'
          title: Ai Overview
        ai_topics:
          items:
            type: string
          type: array
          title: Ai Topics
        created_at:
          type: string
          format: date-time
          title: Created At
        role:
          $ref: '#/components/schemas/SpaceRole'
        memory_count:
          type: integer
          title: Memory Count
        member_count:
          type: integer
          title: Member Count
        connection_count:
          anyOf:
          - type: integer
          - type: 'null'
          title: Connection Count
        items:
          items:
            $ref: '#/components/schemas/VaultItemRead'
          type: array
          title: Items
        members:
          items:
            $ref: '#/components/schemas/SpaceMemberRead'
          type: array
          title: Members
      type: object
      required:
      - id
      - name
      - slug
      - description
      - visibility
      - icon
      - emoji
      - accent
      - pinned
      - ai_overview
      - ai_topics
      - created_at
      - role
      - memory_count
      - member_count
      title: SpaceDetail
    SpaceMemberRead:
      properties:
        user_id:
          type: string
          format: uuid
          title: User Id
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        avatar_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Avatar Url
        role:
          $ref: '#/components/schemas/SpaceRole'
      type: object
      required:
      - user_id
      - name
      - avatar_url
      - role
      title: SpaceMemberRead
    SpaceRead:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        name:
          type: string
          title: Name
        slug:
          type: string
          title: Slug
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        visibility:
          $ref: '#/components/schemas/Visibility'
        icon:
          anyOf:
          - type: string
          - type: 'null'
          title: Icon
        emoji:
          anyOf:
          - type: string
          - type: 'null'
          title: Emoji
        accent:
          anyOf:
          - type: string
          - type: 'null'
          title: Accent
        pinned:
          type: boolean
          title: Pinned
        ai_overview:
          anyOf:
          - type: string
          - type: 'null'
          title: Ai Overview
        ai_topics:
          items:
            type: string
          type: array
          title: Ai Topics
        created_at:
          type: string
          format: date-time
          title: Created At
        role:
          $ref: '#/components/schemas/SpaceRole'
        memory_count:
          type: integer
          title: Memory Count
        member_count:
          type: integer
          title: Member Count
        connection_count:
          anyOf:
          - type: integer
          - type: 'null'
          title: Connection Count
      type: object
      required:
      - id
      - name
      - slug
      - description
      - visibility
      - icon
      - emoji
      - accent
      - pinned
      - ai_overview
      - ai_topics
      - created_at
      - role
      - memory_count
      - member_count
      title: SpaceRead
...
    UpdateSpaceRequest:
      properties:
        name:
          anyOf:
          - type: string
            maxLength: 255
            minLength: 1
          - type: 'null'
          title: Name
        description:
          anyOf:
          - type: string
            maxLength: 2000
          - type: 'null'
          title: Description
        visibility:
          anyOf:
          - $ref: '#/components/schemas/Visibility'
          - type: 'null'
        icon:
          anyOf:
          - type: string
            maxLength: 48
          - type: 'null'
          title: Icon
        emoji:
          anyOf:
          - type: string
            maxLength: 16
          - type: 'null'
          title: Emoji
        accent:
          anyOf:
          - type: string
            maxLength: 24
          - type: 'null'
          title: Accent
        pinned:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Pinned
      type: object
      title: UpdateSpaceRequest
      description: 'Every field optional, and every field one a caller is allowed to set.


        `None` means "leave it alone" rather than "clear it", except for `icon`, `emoji` and

        `accent` where the empty string is the way to clear -- a nullable field with two

        meanings for null is a field nobody can use correctly. That is also why `icon` carries

        no pattern here: `""` has to be accepted, and a pattern that allows the empty string

        allows it everywhere. `_validate_icon` does the check instead.'
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: recall_session
      description: Signed JWT session cookie set by /api/v1/auth/google/callback.